+++ server/streamreader/stream_manager.cpp
@@ -22,6 +22,7 @@
 // local headers
 #include "airplay_stream.hpp"
 #include "common/aixlog.hpp"
+#include <algorithm>
 #ifdef HAS_ALSA
 #include "alsa_stream.hpp"
 #endif
@@ -78,6 +79,13 @@ PcmStreamPtr StreamManager::addStream(StreamUri& streamUri, PcmStream::Source so
     if (streamUri.query.find(kUriChunkMs) == streamUri.query.end())
         streamUri.query[kUriChunkMs] = cpt::to_string(settings_.stream.streamChunkMs);
 
+    auto name = streamUri.query[kUriName];
+    if (name.empty())
+        throw SnapException("Stream name must not be empty");
+
+    auto iter = find_if(streams_.begin(), streams_.end(), [&name](const PcmStreamPtr& stream) { return stream->getName() == name; });
+    if (iter != streams_.end())
+        throw SnapException("Stream with name '" + name + "' already exists");
     //	LOG(DEBUG) << "\nURI: " << streamUri.uri << "\nscheme: " << streamUri.scheme << "\nhost: "
     //		<< streamUri.host << "\npath: " << streamUri.path << "\nfragment: " << streamUri.fragment << "\n";
 
@@ -154,20 +162,13 @@ PcmStreamPtr StreamManager::addStream(StreamUri& streamUri, PcmStream::Source so
     }
 
     if (stream)
-    {
-        for (const auto& s : streams_)
-        {
-            if (s->getName() == stream->getName())
-                throw SnapException("Stream with name \"" + stream->getName() + "\" already exists");
-        }
         streams_.push_back(stream);
-    }
 
     return stream;
 }
 
 
-void StreamManager::removeStream(const std::string& name)
+bool StreamManager::removeStream(const std::string& name)
 {
     LOG(INFO, LOG_TAG) << "Removing stream '" << name << "'\n";
     auto iter = std::find_if(streams_.begin(), streams_.end(), [&name](const PcmStreamPtr& stream) { return stream->getName() == name; });
@@ -176,6 +177,12 @@ void StreamManager::removeStream(const std::string& name)
         (*iter)->stop();
         streams_.erase(iter);
         LOG(DEBUG, LOG_TAG) << "Found and removed stream '" << (*iter)->getName() << "'\n";
+        return true;
+    }
+    else
+    {
+        LOG(WARNING, LOG_TAG) << "Stream '" << name << "' not found\n";
+        return false;
     }
 }
 
@@ -191,9 +198,8 @@ const PcmStreamPtr StreamManager::getDefaultStream() const
     if (streams_.empty())
         return nullptr;
 
-    bool hasDefaultSource = !settings_.stream.defaultSource.empty();
+    auto& default_source = settings_.stream.default_source;
     PcmStreamPtr firstValidStream = nullptr;
-
     for (const auto& stream : streams_)
     {
         if (stream->getCodec() != "null")
@@ -201,7 +207,7 @@ const PcmStreamPtr StreamManager::getDefaultStream() const
             if (firstValidStream == nullptr)
                 firstValidStream = stream;
 
-            if (!hasDefaultSource || (hasDefaultSource && stream->getName() == settings_.stream.defaultSource))
+            if (!default_source.has_value() || (default_source.value() == stream->getName()))
                 return stream;
         }
     }